Start splitting Graphics.subs.php into classes - extracted the "Image" class as a wrapper for the two Imagick and Gd2 classes that do the actual work (in most of the cases) - side effect: the global $gd2 has been removed
