+++ sources/subs/Graphics.subs.php
@@ -22,7 +22,6 @@
 /**
  * Create a thumbnail of the given source.
  *
- * @uses resizeImageFile() function to achieve the resize.
  * @package Graphics
  *
  * @param string $source The name of the source image
@@ -35,17 +34,19 @@ function createThumbnail($source, $max_width, $max_height)
 {
 	global $modSettings;
 
-	$destName = $source . '_thumb.tmp';
+	$destName = $source . '_thumb';
 
 	// Do the actual resize, thumbnails by default strip EXIF data to save space
 	$format = !empty($modSettings['attachment_thumb_png']) ? 3 : 0;
-	$success = resizeImageFile($source, $destName, $max_width, $max_height, $format, true);
+
+	$image = new \ElkArte\Graphics\Image($destName . '.tmp');
+	$success = $image->resizeImageFile($source, $max_width, $max_height, $format, true);
 
 	// Okay, we're done with the temporary stuff.
-	$destName = substr($destName, 0, -4);
-
 	if ($success && @rename($destName . '.tmp', $destName))
+	{
 		return true;
+	}
 	else
 	{
 		@unlink($destName . '.tmp');
@@ -77,18 +78,23 @@ function createThumbnail($source, $max_width, $max_height)
  */
 function reencodeImage($fileName, $preferred_format = 0)
 {
-	if (!resizeImageFile($fileName, $fileName . '.tmp', null, null, $preferred_format))
+	$image = new \ElkArte\Graphics\Image($fileName . '.tmp', null, null, $preferred_format);
+	if ($image->resizeImageFile($fileName) === false)
 	{
 		if (file_exists($fileName . '.tmp'))
+		{
 			unlink($fileName . '.tmp');
+		}
 
 		return false;
 	}
 
 	if (!unlink($fileName))
+	{
 		return false;
+	}
 
-	return rename($fileName . '.tmp', $fileName);
+	return $image->move($fileName);
 }
 
 /**
@@ -150,845 +156,6 @@ function checkImageContents($fileName, $extensiveCheck = false)
 }
 
 /**
- * Sets a global $gd2 variable needed by some functions to determine
- * whether the GD2 library is present.
- *
- * @package Graphics
- *
- * @return bool Whether or not GD is available.
- */
-function checkGD()
-{
-	global $gd2;
-
-	// Check to see if GD is installed and what version.
-	if (($extensionFunctions = get_extension_funcs('gd')) === false)
-		return false;
-
-	// Also determine if GD2 is installed and store it in a global.
-	$gd2 = in_array('imagecreatetruecolor', $extensionFunctions) && function_exists('imagecreatetruecolor');
-
-	return true;
-}
-
-/**
- * Checks whether the Imagick class is present.
- *
- * @package Graphics
- *
- * @return bool Whether or not the Imagick extension is available.
- */
-function checkImagick()
-{
-	return class_exists('Imagick', false);
-}
-
-/**
- * See if we have enough memory to thumbnail an image
- *
- * @package Graphics
- * @param int[] $sizes image size
- *
- * @return bool Whether or not the memory is available.
- */
-function imageMemoryCheck($sizes)
-{
-	global $modSettings;
-
-	// Just to be sure
-	if (!is_array($sizes) || $sizes[0] === -1)
-	{
-		return true;
-	}
-
-	// Doing the old 'set it and hope' way?
-	if (empty($modSettings['attachment_thumb_memory']))
-	{
-		detectServer()->setMemoryLimit('128M');
-		return true;
-	}
-
-	// Determine the memory requirements for this image, note: if you want to use an image formula
-	// W x H x bits/8 x channels x Overhead factor
-	// You will need to account for single bit images as GD expands them to an 8 bit and will greatly
-	// overun the calculated value.
-	// The 5 below is simply a shortcut of 8bpp, 3 channels, 1.66 overhead
-	$needed_memory = ($sizes[0] * $sizes[1] * 5);
-
-	// If we need more, lets try to get it
-	return detectServer()->setMemoryLimit($needed_memory, true);
-}
-
-/**
- * Resize an image from a remote location or a local file.
- *
- * What it does:
- *
- * - Puts the resized image at the destination location.
- * - The file would have the format preferred_format if possible,
- * otherwise the default format is jpeg.
- *
- * @package Graphics
- *
- * @param string $source The name of the source image
- * @param string $destination The name of the destination image
- * @param int $max_width The maximum allowed width
- * @param int $max_height The maximum allowed height
- * @param int $preferred_format Used by Imagick/resizeImage
- * @param bool $force_resize Always resize the image (force scale up)
- * @param bool $strip Allow IM to remove exif data as GD always will
- *
- * @return boolean Whether the thumbnail creation was successful.
- */
-function resizeImageFile($source, $destination, $max_width, $max_height, $preferred_format = 0, $strip = false, $force_resize = true)
-{
-	// Nothing to do without GD or IM
-	if (!checkGD() && !checkImagick())
-		return false;
-
-	if (!file_exists($source) && substr($source, 0, 7) !== 'http://' && substr($source, 0, 8) !== 'https://')
-	{
-		return false;
-	}
-
-	static $default_formats = array(
-		'1' => 'gif',
-		'2' => 'jpeg',
-		'3' => 'png',
-		'6' => 'bmp',
-		'15' => 'wbmp'
-	);
-
-	require_once(SUBSDIR . '/Package.subs.php');
-	require_once(SUBSDIR . '/Attachments.subs.php');
-
-	// Get the image file, we have to work with something after all
-	$fp_destination = fopen($destination, 'wb');
-	if ($fp_destination && (substr($source, 0, 7) === 'http://' || substr($source, 0, 8) === 'https://'))
-	{
-		$fileContents = fetch_web_data($source);
-
-		fwrite($fp_destination, $fileContents);
-		fclose($fp_destination);
-
-		$sizes = elk_getimagesize($destination);
-	}
-	elseif ($fp_destination)
-	{
-		$sizes = elk_getimagesize($source);
-
-		$fp_source = fopen($source, 'rb');
-		if ($fp_source !== false)
-		{
-			while (!feof($fp_source))
-			{
-				fwrite($fp_destination, fread($fp_source, 8192));
-			}
-			fclose($fp_source);
-		}
-		else
-			$sizes = array(-1, -1, -1);
-
-		fclose($fp_destination);
-	}
-	// We can't get to the file.
-	else
-		$sizes = array(-1, -1, -1);
-
-	// See if we have -or- can get the needed memory for this operation
-	if (checkGD() && !imageMemoryCheck($sizes))
-		return false;
-
-	// A known and supported format?
-	if (checkImagick() && isset($default_formats[$sizes[2]]))
-	{
-		return resizeImage(null, $destination, null, null, $max_width, $max_height, $force_resize, $preferred_format, $strip);
-	}
-	elseif (checkGD() && isset($default_formats[$sizes[2]]) && function_exists('imagecreatefrom' . $default_formats[$sizes[2]]))
-	{
-		$imagecreatefrom = 'imagecreatefrom' . $default_formats[$sizes[2]];
-		if ($src_img = @$imagecreatefrom($destination))
-		{
-			return resizeImage($src_img, $destination, imagesx($src_img), imagesy($src_img), $max_width === null ? imagesx($src_img) : $max_width, $max_height === null ? imagesy($src_img) : $max_height, $force_resize, $preferred_format, $strip);
-		}
-	}
-
-	return false;
-}
-
-/**
- * Resize an image proportionally to fit within the defined max_width and max_height limits
- *
- * What it does:
- *
- * - Will do nothing to the image if the file fits within the size limits
- * - If Image Magick is present it will use those function over any GD solutions
- * - If GD2 is present, it'll use it to achieve better quality (imagecopyresampled)
- * - Saves the new image to destination_filename, in the preferred_format
- * if possible, default is jpeg.
- *
- * @uses GD
- * @uses Imagick
- *
- * @package Graphics
- * @param resource|null $src_img null for Imagick images, resource form imagecreatefrom for GD
- * @param string $destName
- * @param int $src_width The width of the source image
- * @param int $src_height The height of the source image
- * @param int $max_width The maximum allowed width
- * @param int $max_height The maximum allowed height
- * @param bool $force_resize = false Whether to override defaults and resize it
- * @param int $preferred_format - The preferred format
- *   - 0 to use jpeg
- *   - 1 for gif
- *   - 2 to force jpeg
- *   - 3 for png
- *   - 6 for bmp
- *   - 15 for wbmp
- * @param bool $strip Whether to have IM strip EXIF data as GD will
- *
- * @return bool Whether resize was successful.
- */
-function resizeImage($src_img, $destName, $src_width, $src_height, $max_width, $max_height, $force_resize = false, $preferred_format = 0, $strip = false)
-{
-	global $gd2;
-
-	if (checkImagick())
-	{
-		// These are the file formats we know about
-		static $default_formats = array(
-			'1' => 'gif',
-			'2' => 'jpeg',
-			'3' => 'png',
-			'6' => 'bmp',
-			'15' => 'wbmp'
-		);
-		$preferred_format = empty($preferred_format) || !isset($default_formats[$preferred_format]) ? 2 : $preferred_format;
-
-		// Since Imagick can throw exceptions, lets catch them
-		try
-		{
-			// Get a new instance of Imagick for use
-			$imagick = new Imagick($destName);
-
-			// Set the input and output image size
-			$src_width = empty($src_width) ? $imagick->getImageWidth() : $src_width;
-			$src_height = empty($src_height) ? $imagick->getImageHeight() : $src_height;
-
-			// The behavior of bestfit changed in Imagick 3.0.0 and it will now scale up, we prevent that
-			$dest_width = empty($max_width) ? $src_width : ($force_resize ? $max_width : min($max_width, $src_width));
-			$dest_height = empty($max_height) ? $src_height : ($force_resize ? $max_height : min($max_height, $src_height));
-
-			// Set jpeg image quality to 80
-			if ($default_formats[$preferred_format] === 'jpeg')
-			{
-				$imagick->borderImage('white', 0, 0);
-				$imagick->setImageCompression(Imagick::COMPRESSION_JPEG);
-				$imagick->setImageCompressionQuality(80);
-			}
-
-			// Create a new image in our preferred format and resize it if needed
-			$imagick->setImageFormat($default_formats[$preferred_format]);
-			$imagick->resizeImage($dest_width, $dest_height, Imagick::FILTER_LANCZOS, 1, true);
-
-			// Remove EXIF / ICC data?
-			if ($strip)
-			{
-				$imagick->stripImage();
-			}
-
-			// Save the new image in the destination location
-			$success = $imagick->writeImage($destName);
-
-			// Free resources associated with the Imagick object
-			$imagick->clear();
-		}
-		catch (Exception $e)
-		{
-			$success = false;
-		}
-
-		return !empty($success);
-	}
-	elseif (checkGD())
-	{
-		$success = false;
-
-		// Determine whether to resize to max width or to max height (depending on the limits.)
-		if (!empty($max_width) || !empty($max_height))
-		{
-			if (!empty($max_width) && (empty($max_height) || $src_height * $max_width / $src_width <= $max_height))
-			{
-				$dst_width = $max_width;
-				$dst_height = floor($src_height * $max_width / $src_width);
-			}
-			elseif (!empty($max_height))
-			{
-				$dst_width = floor($src_width * $max_height / $src_height);
-				$dst_height = $max_height;
-			}
-
-			// Don't bother resizing if it's already smaller...
-			if (!empty($dst_width) && !empty($dst_height) && ($dst_width < $src_width || $dst_height < $src_height || $force_resize))
-			{
-				// (make a true color image, because it just looks better for resizing.)
-				if ($gd2)
-				{
-					$dst_img = imagecreatetruecolor($dst_width, $dst_height);
-
-					// Deal nicely with a PNG - because we can.
-					if ((!empty($preferred_format)) && ($preferred_format == 3))
-					{
-						imagealphablending($dst_img, false);
-						if (function_exists('imagesavealpha'))
-							imagesavealpha($dst_img, true);
-					}
-				}
-				else
-					$dst_img = imagecreate($dst_width, $dst_height);
-
-				// Resize it!
-				if ($gd2)
-					imagecopyresampled($dst_img, $src_img, 0, 0, 0, 0, $dst_width, $dst_height, $src_width, $src_height);
-				else
-					imagecopyresamplebicubic($dst_img, $src_img, 0, 0, 0, 0, $dst_width, $dst_height, $src_width, $src_height);
-			}
-			else
-				$dst_img = $src_img;
-		}
-		else
-			$dst_img = $src_img;
-
-		// Save the image as ...
-		if (!empty($preferred_format) && ($preferred_format == 3) && function_exists('imagepng'))
-			$success = imagepng($dst_img, $destName);
-		elseif (!empty($preferred_format) && ($preferred_format == 1) && function_exists('imagegif'))
-			$success = imagegif($dst_img, $destName);
-		elseif (function_exists('imagejpeg'))
-			$success = imagejpeg($dst_img, $destName, 80);
-
-		// Free the memory.
-		imagedestroy($src_img);
-		if ($dst_img != $src_img)
-			imagedestroy($dst_img);
-
-		return $success;
-	}
-	// Without Imagick or GD, no image resizing at all.
-	else
-		return false;
-}
-
-/**
- * Calls GD or ImageMagick functions to correct an images orientation
- * based on the EXIF orientation flag
- *
- * @param string $image_name
- */
-function autoRotateImage($image_name)
-{
-	if (checkImagick())
-	{
-		autoRotateImageWithIM($image_name);
-	}
-	elseif (checkGD())
-	{
-		autoRotateImageWithGD($image_name);
-	}
-}
-
-/**
- * Autorotate an image based on its EXIF Orientation tag.
- *
- * What it does:
- *
- * - GD only
- * - Checks exif data for orientation flag and rotates image so its proper
- * - Does not update orientation flag as GD removes EXIF data
- * - Only works with jpeg images, could add TIFF as well
- * - Writes the update image back to $image_name
- *
- * @package Graphics
- * @uses GD
- *
- * @param string $image_name full location of the file
- *
- * @return bool
- */
-function autoRotateImageWithGD($image_name)
-{
-	// Read the EXIF data
-	$exif = function_exists('exif_read_data') ? @exif_read_data($image_name) : array();
-
-	// We're only interested in the exif orientation
-	$orientation = isset($exif['Orientation']) ? $exif['Orientation'] : 0;
-
-	// For now we only process jpeg images, so check that we have one
-	$sizes = elk_getimagesize($image_name);
-
-	// Not a jpeg or not rotated, done!
-	if ($sizes[2] !== 2 || $orientation === 0 || !imageMemoryCheck($sizes))
-	{
-		return false;
-	}
-
-	// Load the image object so we can begin the transformation(s)
-	$source = imagecreatefromjpeg($image_name);
-
-	// Time to spin and mirror as needed
-	switch ($orientation)
-	{
-		// 0 & 1 Not set or Normal
-		case 0:
-		case 1:
-			break;
-		// 2 Mirror image, Normal orientation
-		case 2:
-			$source = flopImageGD($source);
-			break;
-		// 3 Normal image, rotated 180
-		case 3:
-			$source = rotateImageGD($source, 180);
-			break;
-		// 4 Mirror image, rotated 180
-		case 4:
-			$source = flipImageGD($source);
-			break;
-		// 5 Mirror image, rotated 90 CCW
-		case 5:
-			$source = flopImageGD($source);
-			$source = rotateImageGD($source, 90);
-			break;
-		// 6 Normal image, rotated 90 CCW
-		case 6:
-			$source = rotateImageGD($source, -90);
-			break;
-		// 7 Mirror image, rotated 90 CW
-		case 7:
-			$source = flopImageGD($source);
-			$source = rotateImageGD($source, -90);
-			break;
-		// 8 Normal image, rotated 90 CW
-		case 8:
-			$source = rotateImageGD($source, 90);
-			break;
-	}
-
-	// Save the updated image, free resources
-	imagejpeg($source, $image_name);
-	imagedestroy($source);
-
-	return true;
-}
-
-/**
- * Autorotate an image based on its EXIF Orientation tag.
- *
- * What it does:
- *
- * - ImageMagick only
- * - Checks exif data for orientation flag and rotates image so its proper
- * - Updates orientation flag if rotation was required
- * - Writes the update image back to $image_name
- *
- * @uses Imagick
- *
- * @param string $image_name
- *
- * @return bool
- */
-function autoRotateImageWithIM($image_name)
-{
-	try
-	{
-		// Get a new instance of Imagick for use
-		$image = new Imagick($image_name);
-
-		// This method should exist if Imagick has been compiled against ImageMagick version
-		// 6.3.0 or higher which is forever ago, but we check anyway ;)
-		if (!method_exists($image, 'getImageOrientation'))
-		{
-			return false;
-		}
-
-		$orientation = $image->getImageOrientation();
-		switch ($orientation)
-		{
-			// 0 & 1 Not set or Normal
-			case Imagick::ORIENTATION_UNDEFINED:
-			case Imagick::ORIENTATION_TOPLEFT:
-				break;
-			// 2 Mirror image, Normal orientation
-			case Imagick::ORIENTATION_TOPRIGHT:
-				$image->flopImage();
-				break;
-			// 3 Normal image, rotated 180
-			case Imagick::ORIENTATION_BOTTOMRIGHT:
-				$image->rotateImage('#000', 180);
-				break;
-			// 4 Mirror image, rotated 180
-			case Imagick::ORIENTATION_BOTTOMLEFT:
-				$image->flipImage();
-				break;
-			// 5 Mirror image, rotated 90 CCW
-			case Imagick::ORIENTATION_LEFTTOP:
-				$image->rotateImage('#000', 90);
-				$image->flopImage();
-				break;
-			// 6 Normal image, rotated 90 CCW
-			case Imagick::ORIENTATION_RIGHTTOP:
-				$image->rotateImage('#000', 90);
-				break;
-			// 7 Mirror image, rotated 90 CW
-			case Imagick::ORIENTATION_RIGHTBOTTOM:
-				$image->rotateImage('#000', -90);
-				$image->flopImage();
-				break;
-			// 8 Normal image, rotated 90 CW
-			case Imagick::ORIENTATION_LEFTBOTTOM:
-				$image->rotateImage('#000', -90);
-				break;
-		}
-
-		// Now that it's auto-rotated, make sure the EXIF data is correctly updated
-		if ($orientation >= 2)
-		{
-			$image->setImageOrientation(Imagick::ORIENTATION_TOPLEFT);
-		}
-
-		// Save the new image in the destination location
-		$success = $image->writeImage($image_name);
-
-		// Free resources associated with the Imagick object
-		$image->clear();
-	}
-	catch (Exception $e)
-	{
-		$success = false;
-	}
-
-	return $success;
-}
-
-/**
- * Rotate an image by X degrees, GD function
- *
- * @param resource $image
- * @param int $degrees
- *
- * @package Graphics
- * @uses GD
- * @return resource
- */
-function rotateImageGD($image, $degrees)
-{
-	// Kind of need this to do anything
-	if (function_exists('imagerotate'))
-	{
-		// Use positive degrees so GD does not get confused
-		$degrees -= floor($degrees / 360) * 360;
-
-		// Rotate away
-		$background = imagecolorallocatealpha($image, 255, 255, 255, 127);
-		$image = imagerotate($image, $degrees, $background);
-	}
-
-	return $image;
-}
-
-/**
- * Flop an image using GD functions by copying top to bottom / flop
- *
- * @param resource $image
- *
- * @return resource
- * @package Graphics
- * @uses GD
- */
-function flopImageGD($image)
-{
-	return flipImageGD($image, 'horizontal');
-}
-
-/**
- * Flip an image using GD function by copying top to bottom / flip vertical
- *
- * @param resource $image
- * @param string $axis vertical for flip about vertical otherwise horizontal flip
- *
- * @return resource
- * @package Graphics
- * @uses GD
- */
-function flipImageGD($image, $axis = 'vertical')
-{
-	imageflip($image, $axis === 'vertical' ? IMG_FLIP_VERTICAL : IMG_FLIP_HORIZONTAL);
-
-	return $image;
-}
-
-/**
- * Copy / resize an image using GD bicubic methods
- *
- * What it does:
- *
- * - Used when imagecopyresample() is not available
- * - Uses bicubic resizing methods which are lower quality then imagecopyresample
- *
- * @package Graphics
- * @param resource $dst_img The destination image - a GD image resource
- * @param resource $src_img The source image - a GD image resource
- * @param int $dst_x The "x" coordinate of the destination image
- * @param int $dst_y The "y" coordinate of the destination image
- * @param int $src_x The "x" coordinate of the source image
- * @param int $src_y The "y" coordinate of the source image
- * @param int $dst_w The width of the destination image
- * @param int $dst_h The height of the destination image
- * @param int $src_w The width of the destination image
- * @param int $src_h The height of the destination image
- */
-function imagecopyresamplebicubic($dst_img, $src_img, $dst_x, $dst_y, $src_x, $src_y, $dst_w, $dst_h, $src_w, $src_h)
-{
-	$palsize = imagecolorstotal($src_img);
-	for ($i = 0; $i < $palsize; $i++)
-	{
-		$colors = imagecolorsforindex($src_img, $i);
-		imagecolorallocate($dst_img, $colors['red'], $colors['green'], $colors['blue']);
-	}
-
-	$scaleX = ($src_w - 1) / $dst_w;
-	$scaleY = ($src_h - 1) / $dst_h;
-
-	$scaleX2 = (int) $scaleX / 2;
-	$scaleY2 = (int) $scaleY / 2;
-
-	for ($j = $src_y; $j < $dst_h; $j++)
-	{
-		$sY = (int) $j * $scaleY;
-		$y13 = $sY + $scaleY2;
-
-		for ($i = $src_x; $i < $dst_w; $i++)
-		{
-			$sX = (int) $i * $scaleX;
-			$x34 = $sX + $scaleX2;
-
-			$color1 = imagecolorsforindex($src_img, imagecolorat($src_img, $sX, $y13));
-			$color2 = imagecolorsforindex($src_img, imagecolorat($src_img, $sX, $sY));
-			$color3 = imagecolorsforindex($src_img, imagecolorat($src_img, $x34, $y13));
-			$color4 = imagecolorsforindex($src_img, imagecolorat($src_img, $x34, $sY));
-
-			$red = ($color1['red'] + $color2['red'] + $color3['red'] + $color4['red']) / 4;
-			$green = ($color1['green'] + $color2['green'] + $color3['green'] + $color4['green']) / 4;
-			$blue = ($color1['blue'] + $color2['blue'] + $color3['blue'] + $color4['blue']) / 4;
-
-			$color = imagecolorresolve($dst_img, $red, $green, $blue);
-			if ($color == -1)
-			{
-				if ($palsize++ < 256)
-					imagecolorallocate($dst_img, $red, $green, $blue);
-				$color = imagecolorclosest($dst_img, $red, $green, $blue);
-			}
-
-			imagesetpixel($dst_img, $i + $dst_x - $src_x, $j + $dst_y - $src_y, $color);
-		}
-	}
-}
-
-if (!function_exists('imagecreatefrombmp'))
-{
-	/**
-	 * It is set only if it doesn't already exist (for forwards compatibility.)
-	 *
-	 * What it does:
-	 *
-	 * - It only supports uncompressed bitmaps.
-	 * - It only supports standard windows bitmaps (no os/2 variants)
-	 * - Returns an image identifier representing the bitmap image
-	 * obtained from the given filename.
-	 *
-	 * @package Graphics
-	 * @param string $filename The name of the file
-	 * @return resource An image identifier representing the bitmap image
-	 */
-	function imagecreatefrombmp($filename)
-	{
-		global $gd2;
-
-		$fp = fopen($filename, 'rb');
-
-		$errors = error_reporting(0);
-
-		// Unpack the general information about the Bitmap Image File, first 14 Bytes
-		$header = unpack('vtype/Vsize/Vreserved/Voffset', fread($fp, 14));
-
-		// Unpack the DIB header, it stores detailed information about the bitmap image the pixel format, 40 Bytes long
-		$info = unpack('Vsize/Vwidth/Vheight/vplanes/vbits/Vcompression/Vimagesize/Vxres/Vyres/Vncolor/Vcolorimportant', fread($fp, 40));
-
-		// Not a standard bitmap, bail out
-		if ($header['type'] != 0x4D42)
-			return false;
-
-		// Create our image canvas with the given WxH
-		if ($gd2)
-			$dst_img = imagecreatetruecolor($info['width'], $info['height']);
-		else
-			$dst_img = imagecreate($info['width'], $info['height']);
-
-		// Color bitCounts 1,4,8 have palette information we use
-		$palette = array();
-		if ($info['bits'] == 1 || $info['bits'] == 4 || $info['bits'] == 8)
-		{
-			$palette_size = $header['offset'] - 54;
-
-			// Read the palette data
-			$palettedata = fread($fp, $palette_size);
-
-			// Create the rgb color array
-			$n = 0;
-			for ($j = 0; $j < $palette_size; $j++)
-			{
-				$b = ord($palettedata[$j++]);
-				$g = ord($palettedata[$j++]);
-				$r = ord($palettedata[$j++]);
-
-				$palette[$n++] = imagecolorallocate($dst_img, $r, $g, $b);
-			}
-		}
-
-		$scan_line_size = ($info['bits'] * $info['width'] + 7) >> 3;
-		$scan_line_align = $scan_line_size & 3 ? 4 - ($scan_line_size & 3) : 0;
-
-		for ($y = 0, $l = $info['height'] - 1; $y < $info['height']; $y++, $l--)
-		{
-			fseek($fp, $header['offset'] + ($scan_line_size + $scan_line_align) * $l);
-			$scan_line = fread($fp, $scan_line_size);
-
-			if (strlen($scan_line) < $scan_line_size)
-				continue;
-
-			// 32 bits per pixel
-			if ($info['bits'] == 32)
-			{
-				$x = 0;
-				for ($j = 0; $j < $scan_line_size; $x++)
-				{
-					$b = ord($scan_line[$j++]);
-					$g = ord($scan_line[$j++]);
-					$r = ord($scan_line[$j++]);
-					$j++;
-
-					$color = imagecolorexact($dst_img, $r, $g, $b);
-					if ($color == -1)
-					{
-						$color = imagecolorallocate($dst_img, $r, $g, $b);
-
-						// Gah!  Out of colors?  Stupid GD 1... try anyhow.
-						if ($color == -1)
-							$color = imagecolorclosest($dst_img, $r, $g, $b);
-					}
-
-					imagesetpixel($dst_img, $x, $y, $color);
-				}
-			}
-			// 24 bits per pixel
-			elseif ($info['bits'] == 24)
-			{
-				$x = 0;
-				for ($j = 0; $j < $scan_line_size; $x++)
-				{
-					$b = ord($scan_line[$j++]);
-					$g = ord($scan_line[$j++]);
-					$r = ord($scan_line[$j++]);
-
-					$color = imagecolorexact($dst_img, $r, $g, $b);
-					if ($color == -1)
-					{
-						$color = imagecolorallocate($dst_img, $r, $g, $b);
-
-						// Gah!  Out of colors?  Stupid GD 1... try anyhow.
-						if ($color == -1)
-							$color = imagecolorclosest($dst_img, $r, $g, $b);
-					}
-
-					imagesetpixel($dst_img, $x, $y, $color);
-				}
-			}
-			// 16 bits per pixel
-			elseif ($info['bits'] == 16)
-			{
-				$x = 0;
-				for ($j = 0; $j < $scan_line_size; $x++)
-				{
-					$b1 = ord($scan_line[$j++]);
-					$b2 = ord($scan_line[$j++]);
-
-					$word = $b2 * 256 + $b1;
-
-					$b = (($word & 31) * 255) / 31;
-					$g = ((($word >> 5) & 31) * 255) / 31;
-					$r = ((($word >> 10) & 31) * 255) / 31;
-
-					// Scale the image colors up properly.
-					$color = imagecolorexact($dst_img, $r, $g, $b);
-					if ($color == -1)
-					{
-						$color = imagecolorallocate($dst_img, $r, $g, $b);
-
-						// Gah!  Out of colors?  Stupid GD 1... try anyhow.
-						if ($color == -1)
-							$color = imagecolorclosest($dst_img, $r, $g, $b);
-					}
-
-					imagesetpixel($dst_img, $x, $y, $color);
-				}
-			}
-			// 8 bits per pixel
-			elseif ($info['bits'] == 8)
-			{
-				$x = 0;
-				for ($j = 0; $j < $scan_line_size; $x++)
-					imagesetpixel($dst_img, $x, $y, $palette[ord($scan_line[$j++])]);
-			}
-			// 4 bits per pixel
-			elseif ($info['bits'] == 4)
-			{
-				$x = 0;
-				for ($j = 0; $j < $scan_line_size; $x++)
-				{
-					$byte = ord($scan_line[$j++]);
-
-					imagesetpixel($dst_img, $x, $y, $palette[(int) ($byte / 16)]);
-					if (++$x < $info['width'])
-						imagesetpixel($dst_img, $x, $y, $palette[$byte & 15]);
-				}
-			}
-			// 1 bit
-			elseif ($info['bits'] == 1)
-			{
-				$x = 0;
-				for ($j = 0; $j < $scan_line_size; $x++)
-				{
-					$byte = ord($scan_line[$j++]);
-
-					imagesetpixel($dst_img, $x, $y, $palette[(($byte) & 128) != 0]);
-					for ($shift = 1; $shift < 8; $shift++)
-					{
-						if (++$x < $info['width'])
-							imagesetpixel($dst_img, $x, $y, $palette[(($byte << $shift) & 128) != 0]);
-					}
-				}
-			}
-		}
-
-		fclose($fp);
-
-		error_reporting($errors);
-
-		return $dst_img;
-	}
-}
-
-/**
  * Show an image containing the visual verification code for registration.
  *
  * What it does:
@@ -1004,12 +171,13 @@ if (!function_exists('imagecreatefrombmp'))
  */
 function showCodeImage($code)
 {
-	global $gd2, $settings, $user_info, $modSettings;
+	global $settings, $user_info, $modSettings;
 
 	// What type are we going to be doing?
 	// Note: The higher the value of visual_verification_type the harder the verification is
 	// from 0 as disabled through to 4 as "Very hard".
 	$imageType = $modSettings['visual_verification_type'];
+	$gd2 = \ElkArte\Graphics\Gd2::canUse(true);
 
 	// Special case to allow the admin center to show samples.
 	if ($user_info['is_admin'] && isset($_GET['type']))
@@ -1381,163 +549,4 @@ function showLetterImage($letter)
 
 	// Nothing more to come.
 	die();
-}
-
-/**
- * Simple function to generate an image containing some text.
- * It uses preferentially Imagick if present, otherwise GD.
- * Font and size are fixed.
- *
- * @package Graphics
- *
- * @param string $text The text the image should contain
- * @param int $width Width of the final image
- * @param int $height Height of the image
- * @param string $format Type of the image (valid types are png, jpeg, gif)
- *
- * @return boolean|resource The image or false if neither Imagick nor GD are found
- */
-function generateTextImage($text, $width = 100, $height = 100, $format = 'png')
-{
-	$valid_formats = array('jpeg', 'png', 'gif');
-	if (!in_array($format, $valid_formats))
-	{
-		$format = 'png';
-	}
-
-	if (checkImagick() === true)
-	{
-		return generateTextImageWithIM($text, $width, $height, $format);
-	}
-	elseif (checkGD() === true)
-	{
-		return generateTextImageWithGD($text, $width, $height, $format);
-	}
-	else
-	{
-		return false;
-	}
-}
-
-/**
- * Simple function to generate an image containing some text.
- * It uses preferentially Imagick if present, otherwise GD.
- * Font and size are fixed.
- *
- * @uses GD
- *
- * @package Graphics
- *
- * @param string $text The text the image should contain
- * @param int $width Width of the final image
- * @param int $height Height of the image
- * @param string $format Type of the image (valid types are png, jpeg, gif)
- *
- * @return resource|boolean The image
- */
-function generateTextImageWithGD($text, $width = 100, $height = 100, $format = 'png')
-{
-	global $settings;
-
-	$create_function = 'image' . $format;
-
-	// Create a white filled box
-	$image = imagecreate($width, $height);
-	imagecolorallocate($image, 255, 255, 255);
-
-	$text_color = imagecolorallocate($image, 0, 0, 0);
-	$font = $settings['default_theme_dir'] . '/fonts/VDS_New.ttf';
-
-	// The loop is to try to fit the text into the image.
-	$true_type = function_exists('imagettftext');
-	$font_size = $true_type ? 28 : 5;
-	do
-	{
-		if ($true_type)
-		{
-			$metric = imagettfbbox($font_size, 0, $font, $text);
-			$text_width = abs($metric[4] - $metric[0]);
-			$text_height = abs($metric[5] - $metric[1]);
-		}
-		else
-		{
-			$text_width = imagefontwidth($font_size) * strlen($text);
-			$text_height = imagefontheight($font_size);
-		}
-	} while ($text_width > $width && $font_size-- > 1);
-
-	$w_offset = ($width - $text_width) / 2;
-	$h_offset = $true_type ? ($height / 2) + ($text_height / 2) : ($height - $text_height) / 2;
-
-	if ($true_type)
-	{
-		imagettftext($image, $font_size, 0, $w_offset, $h_offset, $text_color, $font, $text);
-	}
-	else
-	{
-		imagestring($image, $font_size, $w_offset, $h_offset, $text, $text_color);
-	}
-
-	// Capture the image string
-	ob_start();
-	$result = $create_function($image);
-	$image = ob_get_contents();
-	ob_end_clean();
-
-	return $result ? $image : false;
-}
-
-/**
- * Function to generate an image containing some text.
- * It uses Imagick, Font and size are fixed to fit within width
- *
- * @uses Imagick
- *
- * @package Graphics
- *
- * @param string $text The text the image should contain
- * @param int $width Width of the final image
- * @param int $height Height of the image
- * @param string $format Type of the image (valid types are png, jpeg, gif)
- *
- * @return boolean|resource The image or false on error
- */
-function generateTextImageWithIM($text, $width = 100, $height = 100, $format = 'png')
-{
-	global $settings;
-
-	try
-	{
-		$image = new Imagick();
-		$image->newImage($width, $height, new ImagickPixel('white'));
-		$image->setImageFormat($format);
-
-		// 28pt is ~2em given default font stack
-		$font_size = 28;
-
-		$draw = new ImagickDraw();
-		$draw->setStrokeColor(new ImagickPixel('#000000'));
-		$draw->setFillColor(new ImagickPixel('#000000'));
-		$draw->setStrokeWidth(0);
-		$draw->setTextAlignment(Imagick::ALIGN_CENTER);
-		$draw->setFont($settings['default_theme_dir'] . '/fonts/VDS_New.ttf');
-
-		// Make sure the text will fit the the allowed space
-		do
-		{
-			$draw->setFontSize($font_size);
-			$metric = $image->queryFontMetrics($draw, $text);
-			$text_width = (int) $metric['textWidth'];
-		} while ($text_width > $width && $font_size-- > 1);
-
-		// Place text in center of block
-		$image->annotateImage($draw, $width / 2, $height / 2 + $font_size / 4, 0, $text);
-		$image = $image->getImageBlob();
-
-		return $image;
-	}
-	catch (Exception $e)
-	{
-		return false;
-	}
 }
